asd
fa

+++ CMP303_Final/Coursework/CMP105App/GraphicsTools.cpp
@@ -128,7 +128,7 @@ void GraphicsTools::postGameElementsRender(sf::RenderWindow* hwnd,float scoreP1,
 {
 	player1Score.setString("Your score: "+std::to_string((int)scoreP1));
 	player2Score.setString("Player2 Score: " + std::to_string((int)scoreP2));
-	gameStartTimerText.setString("Game time: " + std::to_string((int)gameTime));
+	gameStartTimerText.setString("Game time: " + std::to_string(gameTime));
 	
 	hwnd->draw(player1Score);
 	hwnd->draw(player2Score);
